stemming and word/ sentence dropout implemented
Added possibility to stem the LRP results to get a result which is less "noisy" in terms of börjninigar på ord.

Added posibility to use a dropout method for both words and sentences during preprocessesing with word filters

Added removal of stop words during creation of word filter. Such that no stop words are used in the filters

Implemented some functions for methods used repeatedly.

Generally cleaned the code a bit